Add files via upload
The data is as a result of a chemical analysis of wines grown in the same region in Italy by three(3) different cultivators. there are fifteen different measurements taken for different constituents found in the three types of wine. Using Random forest and Support Vector machine as the Machine models or Algorithms.